requirements.txt
google-adk>=0.1.0
requests
prometheus-client
flask
pytest
python-dotenv
